Update Main Code
This is the simplified version, and I believe the last version I'll release. This program just wasn't as fun as I imagined, and I don't know why I decided to do this in the first place, since I do not like text-based games hahaha. I'll instead work on another project.